+++ Extensions/Ansible/Src/Tasks/Ansible/package.json
@@ -8,9 +8,8 @@
   "dependencies": {
     "scp2": "^0.5.0",
     "ssh2": "^0.5.5",
-    "abab": "^1.0.3",
-    "vso-node-api": "6.0.1-preview",
-    "vsts-task-lib" : "^0.9.20"
+    "vso-node-api": "6.0.1-preview ",
+    "vsts-task-lib" : "2.0.5"
   },
   "devDependencies": {},
   "scripts": {
